+++ NEWS.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * New StaticArrays.jl specializations for multiple functions, including:
-  * `exp` on the orthogonal and special orthogonal group in 2 and 3 dimensions.
+  * `exp` and `log` on the orthogonal and special orthogonal group in 2 and 3 dimensions.
   * `get_coordinates` and `get_vector` on the orthogonal and special orthogonal group in 2 and 3 dimensions, for `LieAlgebraOrthogonalBasis`.
 * More generic implementation of non-mutating `get_vector_lie` on `AbstractProductGroupOperation` groups.